pyproject.toml
[project]
name = "brownfield-cartographer"
version = "0.1.0"
description = "Multi-agent codebase intelligence system for rapid FDE onboarding in production environments"
requires-python = ">=3.11,<3.13"
dependencies = [
    "tree-sitter==0.21.3",
    "tree-sitter-languages==1.10.2",
    "sqlglot>=23.0.0",
    "networkx>=3.2",
    "pydantic>=2.5",
    "langgraph>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "chromadb>=0.5.0",
    "streamlit>=1.35.0",
    "pyvis>=0.3.2",
    "plotly>=5.20.0",
    "gitpython>=3.1.40",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "PyYAML>=6.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
semantic = [
    "sentence-transformers>=2.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
cartographer = "src.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
